Organised more into classes (Error in this version)
