+++ Autodocking 2/Autodocking 2.csproj
@@ -52,6 +52,7 @@
     <Compile Include="HomeLocation.cs" />
     <Compile Include="IOHandler.cs" />
     <Compile Include="GlobalSuppressions.cs" />
+    <Compile Include="PID.cs" />
     <Compile Include="ShipSystemsAnalyzer.cs" />
     <Compile Include="MDK\Bootstrapper.cs" />
     <AdditionalFiles Include="MDK\MDK.options.props">
