+++ Autodocking 2/Instructions.readme
@@ -1,5 +1,10 @@
-﻿R e a d m e
------------
+﻿
 
-In this file you can include any instructions or other comments you want to have injected onto the 
-top of your final script. You can safely delete this file if you do not want any such comments.
+/////////////////////////////////////////////////////
+    Auto Docking 2.0
+    
+    Author:  Spug
+    Please leave credit if the ship is
+    used on the workshop.
+    (Extra thanks to Whip for his PID controller)
+/////////////////////////////////////////////////////
